TT: Add basic CI
The version of glib on RHEL8 is too old for modern qemu so we can't do
builds there.

Adds a basic build of the riscv targets.

Build on Ubuntu 22.04 because some people use that for development
containers. Add 25.10 as it's the latest release, and Debian because
that's what I run on my macbook.