target/riscv/debug: Put mcontrol load/store match address into tval
When an mc/mc6 trigger triggers on a load/store and raises a breakpoint
exception, put the matching memory address into tval, as is already
done for exec addresses matching.

Tval for breakpoint exceptions is optional in the spec, but it seems
more consistent to implement it for load/store if exec is implemented.

Signed-off-by: Nicholas Piggin <npiggin@gmail.com>